+++ lib/redox/models/visit.rb
@@ -16,9 +16,12 @@ module Redox
       property :Type, from: :type, required: false
       property :Reason, from: :reason, required: false
       property :Equipment, from: :equipment, required: false
+      property :CancelReason, from: :cancel_reason, required: false
+      property :DischargeDateTime, from: :end, required: false
 
       alias_method :insurances, :Insurances
       alias_method :start, :VisitDateTime
+      alias_method :end, :DischargeDateTime
 
       def department=(v)
         self[:Location] ||= DEFAULT_LOCATION
@@ -45,7 +48,7 @@ module Redox
       def to_h
         result = super.to_h
 
-        %w[VisitDateTime].each do |k|
+        %w[VisitDateTime DischargeDateTime].each do |k|
           result[key][k] = Redox::Models.format_datetime(result[key][k])
         end