+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2021-02-03
+### Added
+- Visit#CancelReason
+- Visit#DischargeDateTime
+
 ## [1.6.0] - 2021-01-20
 ### Added
 - Media model
@@ -47,12 +52,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Visit#as\_json
 
 ## [1.3.0] - 2020-09-17
-### Added 
+### Added
 - Transaction model
 - Financial#create
 
 ## [1.2.0] - 2020-07-30
-### Added 
+### Added
 - Visit model
 - Model#insurances helper for Patient.Insurances || Visit.Insurances